fix bug in nbap watt reports
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@4096 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.awe.neighbours/src/org/amanzi/awe/neighbours/gpeh/NBAPWattExportProvider.java
@@ -101,7 +101,7 @@ public class NBAPWattExportProvider extends ExportProvider3GPP {
         if (power == null) {
             return;
         }
-        double maxTrPowWatt = Math.pow(10, -3) * Math.pow(10, power / 100);
+        double maxTrPowWatt = Math.pow(10, -3) * Math.pow(10, (double)power / 100);
             for (int i = value3gpp.getMin3GPP(); i <= value3gpp.getMax3GPP(); i++) {
                 int txpower = (int)Math.ceil(maxTrPowWatt * value3gpp.getRightBound(i)/ 1000);// (txpower=TxPower                                                                           // i - 0-1000
                 //analyse txower*10
